Refactored boilerplate
Moved most of the code from provider.go to the resource itself (in this case the randomresource).

Made the provider.go code more abstract so the code that does the work lives in the resource file(s).

Added config code to get config from the stack config or env var

Added unknown resource which just returns errors for all the functions if the resource name/type isn't valid