Adding ci-mgmt (#177)
This updates the boilerplate to use workflows managed by ci-mgmt.

* The `Makefile` was largely copied from docker-build and command providers.
* Workflows were generated from this branch of ci-mgmt https://github.com/pulumi/ci-mgmt/pull/1549

The most notable change is the provider is now referred to as "provider-boilerplate" throughout, instead of "xyz". There are assumptions that break between ci-mgmt and the provider when the repo's suffix doesn't match the name in the schema, and it's simply not worth spending time to fix that right now.

Fixes https://github.com/pulumi/home/issues/3745