Fix SDK build (#194)
One problem with copying mark targets with pulumi-command is that
ci-mgmt has a ton of exceptions baked in just for that provider. So the
targets don't actually work for other providers.


https://github.com/pulumi/ci-mgmt/blob/master/provider-ci/internal/pkg/templates/native/.github/workflows/build.yml#L327-L333

Command builds its SDK as part of the generate step and no-ops the build
steps. This PR fixes the boilerplate to work with existing ci-mgmt
workflows without needing special-casing.

Fixes https://github.com/pulumi/pulumi-provider-boilerplate/issues/189