Adding back prepare target (#271)
We accidentally clobbered this target a couple months ago. I am unhappy
with needing a branch for MAC vs other in the makefile, but I did a
little bit of light research and to properly to support that
functionality in a single code block while still using SED would be
substantial.

Tested locally, used grep to check that all boilerplate strings were
properly replaced.